feat: create base functionality
the design should be adaptive or responsive from. It is acceptable to change the appearance for the mobile version.
the default size of the frame is 10x10 with 10 mines.
the player should be able to click on cells to reveal them. If the cell contains a mine, the game is over. If the cell does not contain a mine, the number of mines in the surrounding cells should be displayed
Need to add: N moves!